% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vector_playing_cards}
\alias{vector_playing_cards}
\title{Vector Playing Cards Image Names}
\format{
\subsection{\code{vector_playing_cards}}{

A vector of length 52
\describe{
\item{vector_playing_cards}{Names of cards for a 52 card deck}
}
}
}
\source{
\url{https://code.google.com/archive/p/vector-playing-cards/}
}
\usage{
vector_playing_cards
}
\description{
Image names for vector playing cards
}
\keyword{datasets}
